+++ hpc-engine/pca_utils.h
@@ -58,6 +58,7 @@ typedef struct
     double init0DeltaT;             // change of temperature in units of eF, default 0.01
     double init0eps;           // epsilon for convergence, default 1.0e-6
     double init0scmix;         // mixing parameter in self-consitent process, default 0.25
+    double init0kc;            // momentum cutoff for uniform solver
     int init0maxiter;          // maximum number of iterations, default 100000
     int init0debug;            // debug level, default=0 (no debug info),
     int init0save;             // save solution to file?, default=0, if 1 then solution is in file 'outprefix'_uniform.solution
@@ -74,13 +75,19 @@ typedef struct
     // static solver parameters
     double energyconveps; // convergence epsilon for energy- fraction of Effg needed to get convergence, default=1.0e-6
     double npartconveps; // convergence epsilon for particle number- fraction of N_tot=(Na+Nb) needed to get convergence, default=1.0e-6
+    double npartconveps_a; // convergence epsilon for particle a
+    double npartconveps_b; // convergence epsilon for particle b
     double linearmixing; // mixing parameter for linear algorithm, default=0.5
     double muchange; // coefficient for changing chemical potential, default=0.5
+    double muchange_a; // coefficient for changing chemical potential, default=0.5
+    double muchange_b; // coefficient for changing chemical potential, default=0.5
     int maxiters; // maximum number of iterations, default=10000
     double temperature; // temperature in units of eF_a, default=0.01
     double referencekF; // value of reference kF used in calculations, if 0.0 then not set (default)
     int spinsymmetry; // impose spin symmetry, default 0 - no
     double mumaxchange; // maximal change of chemical potential per iteration, in units of Fermi energy, default 0.1
+    double mumaxchange_a; // maximal change of chemical potential per iteration, in units of Fermi energy, default 0.1
+    double mumaxchange_b; // maximal change of chemical potential per iteration, in units of Fermi energy, default 0.1
     int resetit; // if 1 set it=0, otherwise continue from value read from checkpoint file, default resetit=1
     int writewf; // if 1 the code will write wave-functions on exit, default writewf=0
     double writeecut; // only states with |E_n/eF|<writeecut will be written, default writeecut=INFINITY
@@ -131,6 +138,14 @@ typedef struct
     char dataformat[8];                 // format of produced files: wdat or npy, default=wdat
     char initialized; // technical variable, indicating that structure is initialized by the input file
     char stdoutfile[MD_CHAR_LGTH]; // technical variable,
+    
+    // CONSERVATION MONITORING
+    double Econservation_start; // # Start to monitor energy conservation from this time*eF, default Econservation_start=1e12 (infinity)
+    double Econservation_stop;  // # Stop to monitor energy conservation at this time*eF, default Econservation_stop=1e12 (infinity)
+    double Econservation_tol;   // # if |[E(t)-E(start_t)]/E(start_t)|>tol then the code will terminate
+    double Nconservation_start; // # Start to monitor total particle number conservation from this time*eF, default Nconservation_start=0 (infinity)
+    double Nconservation_stop;  // # Stop to monitor total particle number conservation at this time*eF, default Nconservation_stop=1e12 (infinity)
+    double Nconservation_tol;  // # if |[N(t)-N(start_t)]/N(start_t)|>tol then the code will terminate
 
     // POTENTIAL PARAMETERS
     double params[MAX_USER_PARAMS]; // double parameters
@@ -217,5 +232,11 @@ void copy_initcheckpoint();
 void copy_reprowftar();
 
 void save_extradata_to_file(size_t size, void *extra_data);
+
+double max_dxdydz();
+
+int init_conservation_of_quantity(int quantity_id, double value);
+int monitor_conservation_of_quantity(int quantity_id, double time, double value, double start_time, double stop_time, double tolerance);
+void print_conservation_of_quantity(int quantity_id, double value, double tolerance);
 
 #endif