+++ templates/okeanos/td/env.sh
@@ -1,0 +1,13 @@
+# 
+# W-SLDA Toolkit
+# Warsaw University of Technology
+#  
+
+# List here modules that need to be loaded before the compilation,
+# or update system variables, examples
+#  
+# module load ...
+# export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:path
+#
+# Path to WSLDA Toolkit
+# export WSLDA=...